CSN: CSN-leap-308
Use 64 bit values in RRR methods.  For 64 bit I/O this reduces the RRR traffic
by about 50%.  FPGA-side STDIO ring traffic is unchanged (32 bits), though RRR
is likely the rate-limiting component.

+++ modules/leap/libraries/platform-services/common-services/stdio-service/stdio-service.h
@@ -134,7 +134,7 @@ class STDIO_SERVER_CLASS: public RRR_SERVER_CLASS,
     void Cleanup();
 
     // FPGA to software request chunks
-    void Req(UINT32 data, UINT8 eom);
+    void Req(UINT64 data, UINT8 eom);
 };
 
 // server stub
